% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terr.R
\name{terr}
\alias{terr}
\title{The terr function is to generate data based on different distributions.}
\usage{
terr(n, p, K, nr, alpha, sigma1, sigma2, beta, dist)
}
\arguments{
\item{n}{is the total number of observations}

\item{p}{is the number of variables}

\item{K}{is the number of subsets}

\item{nr}{is the number of outliers}

\item{alpha}{is the significance level}

\item{sigma1}{is a parameter related to the distribution}

\item{sigma2}{is a parameter related to the distribution}

\item{beta}{is a coefficient vector}

\item{dist}{is the type of distribution}
}
\value{
A list containing the design matrix X and the response vector y
}
\description{
The terr function is to generate data based on different distributions.
}
\examples{
# Example usage of the terr function
n <- 100
p <- 5
K <- 3
nr <- 10
alpha <- 0.05
sigma1 <- 2
sigma2 <- 5
beta <- c(1, 2, 3, 4, 5)
dist <- "Student t Distribution"
result <- terr(n, p, K, nr, alpha, sigma1, sigma2, beta, dist)
X <- result$X
y <- result$y
}
